requirements.txt
python-dotenv
requests
langchain-google-genai
pydantic
langchain
faiss-cpu
langgraph